[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17708 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,8 +3,8 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
-        int a = input.nextInt();
-        int b = input.nextInt();
+        long a = input.nextInt();
+        long b = input.nextInt();
 
         System.out.println((a + b) * (a - b));
     }
